requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
deprecation==2.1.0
gotrue==2.9.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.25.2
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==0.15.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
realtime==1.0.6
six==1.17.0
sniffio==1.3.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.4
supafunc==0.3.3
typing-inspection==0.4.1
typing_extensions==4.14.1
websockets==12.0
